backend/python/requirements.txt
# PDF処理・Excel編集・PDF生成に必要なPythonパッケージ

# PDF解析
pdfplumber==0.11.0

# Excel編集
openpyxl==3.1.2

# 日付処理
python-dateutil==2.8.2
